New Module Styles, Drop Area has sole file
Made the drop area be in its own component file. Grading dialog will have other things to handle.

New styles for the modules view
